+++ nArchitecture.sln
@@ -23,17 +23,17 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Core.Persistence", "src\cor
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Core.Security", "src\corePackages\Core.Security\Core.Security.csproj", "{00AA7CFA-B348-4303-B3D0-40750DA10B94}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "templateProject", "templateProject", "{11C509DD-21AC-415F-B2F2-EB58985F2703}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "rentACar", "rentACar", "{73336431-4FD3-4BEC-9C39-572BF8AF92EA}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Application", "src\demoProjects\templateProject\Application\Application.csproj", "{78EA22FC-B933-4349-A097-556A076009F1}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Application", "src\demoProjects\rentACar\Application\Application.csproj", "{4F697DAF-A862-45AA-8468-E24D783BA4FA}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Domain", "src\demoProjects\templateProject\Domain\Domain.csproj", "{D0D4D281-24C1-4EE0-88D6-01FA18B90442}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Domain", "src\demoProjects\rentACar\Domain\Domain.csproj", "{5004F1F6-89C3-4685-9D60-9A5F3B1A01D9}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Infrastructure", "src\demoProjects\templateProject\Infrastructure\Infrastructure.csproj", "{1A88B2CE-A954-4709-B15E-F6F1210AD2FD}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Infrastructure", "src\demoProjects\rentACar\Infrastructure\Infrastructure.csproj", "{2C8AA225-63AA-4AC1-B2D1-EB9AB284850A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Persistence", "src\demoProjects\templateProject\Persistence\Persistence.csproj", "{53D6E924-1C32-4039-8192-CE1BEA129C61}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Persistance", "src\demoProjects\rentACar\Persistance\Persistance.csproj", "{85ED318E-9A03-4563-9045-A7652D08963B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebAPI", "src\demoProjects\templateProject\WebAPI\WebAPI.csproj", "{54568B1F-31E7-4BA3-A2ED-826E85050603}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebAPI", "src\demoProjects\rentACar\WebAPI\WebAPI.csproj", "{DE8EF7AC-1AF7-4B9E-A299-8D4A71C26B52}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -65,26 +65,26 @@ Global
 		{00AA7CFA-B348-4303-B3D0-40750DA10B94}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{00AA7CFA-B348-4303-B3D0-40750DA10B94}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{00AA7CFA-B348-4303-B3D0-40750DA10B94}.Release|Any CPU.Build.0 = Release|Any CPU
-		{78EA22FC-B933-4349-A097-556A076009F1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{78EA22FC-B933-4349-A097-556A076009F1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{78EA22FC-B933-4349-A097-556A076009F1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{78EA22FC-B933-4349-A097-556A076009F1}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D0D4D281-24C1-4EE0-88D6-01FA18B90442}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D0D4D281-24C1-4EE0-88D6-01FA18B90442}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D0D4D281-24C1-4EE0-88D6-01FA18B90442}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D0D4D281-24C1-4EE0-88D6-01FA18B90442}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1A88B2CE-A954-4709-B15E-F6F1210AD2FD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1A88B2CE-A954-4709-B15E-F6F1210AD2FD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1A88B2CE-A954-4709-B15E-F6F1210AD2FD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1A88B2CE-A954-4709-B15E-F6F1210AD2FD}.Release|Any CPU.Build.0 = Release|Any CPU
-		{53D6E924-1C32-4039-8192-CE1BEA129C61}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{53D6E924-1C32-4039-8192-CE1BEA129C61}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{53D6E924-1C32-4039-8192-CE1BEA129C61}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{53D6E924-1C32-4039-8192-CE1BEA129C61}.Release|Any CPU.Build.0 = Release|Any CPU
-		{54568B1F-31E7-4BA3-A2ED-826E85050603}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{54568B1F-31E7-4BA3-A2ED-826E85050603}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{54568B1F-31E7-4BA3-A2ED-826E85050603}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{54568B1F-31E7-4BA3-A2ED-826E85050603}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4F697DAF-A862-45AA-8468-E24D783BA4FA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4F697DAF-A862-45AA-8468-E24D783BA4FA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4F697DAF-A862-45AA-8468-E24D783BA4FA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4F697DAF-A862-45AA-8468-E24D783BA4FA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5004F1F6-89C3-4685-9D60-9A5F3B1A01D9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5004F1F6-89C3-4685-9D60-9A5F3B1A01D9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5004F1F6-89C3-4685-9D60-9A5F3B1A01D9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5004F1F6-89C3-4685-9D60-9A5F3B1A01D9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2C8AA225-63AA-4AC1-B2D1-EB9AB284850A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2C8AA225-63AA-4AC1-B2D1-EB9AB284850A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2C8AA225-63AA-4AC1-B2D1-EB9AB284850A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2C8AA225-63AA-4AC1-B2D1-EB9AB284850A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{85ED318E-9A03-4563-9045-A7652D08963B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{85ED318E-9A03-4563-9045-A7652D08963B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{85ED318E-9A03-4563-9045-A7652D08963B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{85ED318E-9A03-4563-9045-A7652D08963B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DE8EF7AC-1AF7-4B9E-A299-8D4A71C26B52}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DE8EF7AC-1AF7-4B9E-A299-8D4A71C26B52}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DE8EF7AC-1AF7-4B9E-A299-8D4A71C26B52}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DE8EF7AC-1AF7-4B9E-A299-8D4A71C26B52}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -98,12 +98,12 @@ Global
 		{4EA615C5-8C04-4EA1-91DA-C86DA54D2E63} = {01B7F227-CE8F-41FD-9AF8-6AF5E5259161}
 		{54635492-7F95-449B-8C5B-493566F68D4A} = {01B7F227-CE8F-41FD-9AF8-6AF5E5259161}
 		{00AA7CFA-B348-4303-B3D0-40750DA10B94} = {01B7F227-CE8F-41FD-9AF8-6AF5E5259161}
-		{11C509DD-21AC-415F-B2F2-EB58985F2703} = {84E05683-0E4A-4FB1-8FD4-9F0C509BC74C}
-		{78EA22FC-B933-4349-A097-556A076009F1} = {11C509DD-21AC-415F-B2F2-EB58985F2703}
-		{D0D4D281-24C1-4EE0-88D6-01FA18B90442} = {11C509DD-21AC-415F-B2F2-EB58985F2703}
-		{1A88B2CE-A954-4709-B15E-F6F1210AD2FD} = {11C509DD-21AC-415F-B2F2-EB58985F2703}
-		{53D6E924-1C32-4039-8192-CE1BEA129C61} = {11C509DD-21AC-415F-B2F2-EB58985F2703}
-		{54568B1F-31E7-4BA3-A2ED-826E85050603} = {11C509DD-21AC-415F-B2F2-EB58985F2703}
+		{73336431-4FD3-4BEC-9C39-572BF8AF92EA} = {84E05683-0E4A-4FB1-8FD4-9F0C509BC74C}
+		{4F697DAF-A862-45AA-8468-E24D783BA4FA} = {73336431-4FD3-4BEC-9C39-572BF8AF92EA}
+		{5004F1F6-89C3-4685-9D60-9A5F3B1A01D9} = {73336431-4FD3-4BEC-9C39-572BF8AF92EA}
+		{2C8AA225-63AA-4AC1-B2D1-EB9AB284850A} = {73336431-4FD3-4BEC-9C39-572BF8AF92EA}
+		{85ED318E-9A03-4563-9045-A7652D08963B} = {73336431-4FD3-4BEC-9C39-572BF8AF92EA}
+		{DE8EF7AC-1AF7-4B9E-A299-8D4A71C26B52} = {73336431-4FD3-4BEC-9C39-572BF8AF92EA}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {C127633B-AB1B-43AE-A516-DCF06F13F757}
