+++ src/demoProjects/rentACar/Application/Services/Repositories/IBrandRepository.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using Core.Persistence.Repositories;
+using Domain.Entities;
+
+namespace Application.Services.Repositories
+{
+    public interface IBrandRepository : IAsyncRepository<Brand>, IRepository<Brand>
+    {
+    }
+}
