+++ src/demoProjects/rentACar/Persistance/Contexts/BaseDbContext.cs
@@ -1,0 +1,43 @@
+﻿using Domain.Entities;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Persistence.Contexts
+{
+    public class BaseDbContext : DbContext
+    {
+        protected IConfiguration Configuration { get; set; }
+        public DbSet<Brand> Brands { get; set; }
+       
+        public BaseDbContext(DbContextOptions dbContextOptions, IConfiguration configuration) : base(dbContextOptions)
+        {
+            Configuration = configuration;
+        }
+
+        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
+        {
+            //if (!optionsBuilder.IsConfigured)
+            //    base.OnConfiguring(
+            //        optionsBuilder.UseSqlServer(Configuration.GetConnectionString("SomeConnectionString")));
+        }
+
+        protected override void OnModelCreating(ModelBuilder modelBuilder)
+        {
+            modelBuilder.Entity<Brand>(a =>
+            {
+                a.ToTable("Brands").HasKey(k => k.Id);
+                a.Property(p => p.Id).HasColumnName("Id");
+                a.Property(p => p.Name).HasColumnName("Name");
+            });
+
+            // Seed data
+            Brand[] brandEntitySeeds = { new(1, "BMW"), new(2, "TOGG") };
+            modelBuilder.Entity<Brand>().HasData(brandEntitySeeds);
+        }
+    }
+}
