+++ src/demoProjects/rentACar/Persistance/PersistanceServiceRegistration.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using Application.Services.Repositories;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Persistance.Repositories;
+using Persistence.Contexts;
+
+namespace Persistance
+{
+    public static class PersistanceServiceRegistration
+    {
+        public static IServiceCollection AddPersistenceServices(this IServiceCollection services,
+                                                                        IConfiguration configuration)
+        {
+            services.AddDbContext<BaseDbContext>(options =>
+                                                     options.UseSqlServer(
+                                                         configuration.GetConnectionString("RentACarConnectionString")));
+            services.AddScoped<IBrandRepository, BrandRepository>();
+
+            return services;
+        }
+    }
+}
+
