+++ src/demoProjects/rentACar/Persistance/Repositories/BrandRepository.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using Application.Services.Repositories;
+using Core.Persistence.Repositories;
+using Domain.Entities;
+using Persistence.Contexts;
+
+namespace Persistance.Repositories
+{
+    public class BrandRepository : EfRepositoryBase<Brand, BaseDbContext>, IBrandRepository
+    {
+        public BrandRepository(BaseDbContext context) : base(context)
+        {
+        }
+    }
+}
+
