+++ src/demoProjects/rentACar/WebAPI/appsettings.json
@@ -1,0 +1,12 @@
+﻿{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  },
+  "ConnectionStrings": {
+    "RentACarConnectionString": "Server=(localdb)\\MSSQLLocalDB;Database=RentACarDb;Trusted_Connection=True;"
+  },
+  "AllowedHosts": "*"
+}
