--- src/demoProjects/templateProject/Application/Features/someFeature/Constants/OperationClaims.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Application.Features.someFeature.Constants
-{
-    public static class OperationClaims
-    {
-        public const string SomeFeatureAdd = "someFeature.add";
-    }
-}
