--- src/demoProjects/templateProject/Application/Features/someFeature/Dtos/CreatedSomeFeatureEntityDto.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Application.Features.someFeature.Dtos
-{
-    public class CreatedSomeFeatureEntityDto
-    {
-        public int Id { get; set; }
-        public string Name { get; set; }
-    }
-}
