--- src/demoProjects/templateProject/Application/Features/someFeature/Profiles/MappingProfiles.cs
@@ -1,16 +1,0 @@
-﻿using AutoMapper;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Application.Features.someFeature.Profiles
-{
-    public class MappingProfiles : Profile
-    {
-        public MappingProfiles()
-        {
-        }
-    }
-}
