--- src/demoProjects/templateProject/Application/Services/Repositories/ISomeFeatureEntityRepository.cs
@@ -1,14 +1,0 @@
-﻿using Core.Persistence.Repositories;
-using Domain.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Application.Services.Repositories
-{
-    public interface ISomeFeatureEntityRepository : IAsyncRepository<SomeFeatureEntity>, IRepository<SomeFeatureEntity>
-    {
-    }
-}
