--- src/demoProjects/templateProject/Infrastructure/InfrastructureServiceRegistration.cs
@@ -1,17 +1,0 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Infrastructure
-{
-    public static class InfrastructureServiceRegistration
-    {
-        public static IServiceCollection AddInfrastructureServices(this IServiceCollection services)
-        {
-            return services;
-        }
-    }
-}
