--- src/demoProjects/templateProject/Persistence/Contexts/BaseDbContext.cs
@@ -1,47 +1,0 @@
-﻿using Domain.Entities;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Persistence.Contexts
-{
-    public class BaseDbContext : DbContext
-    {
-        protected IConfiguration Configuration { get; set; }
-        public DbSet<SomeFeatureEntity> SomeFeatureEntities { get; set; }
-       
-
-        public BaseDbContext(DbContextOptions dbContextOptions, IConfiguration configuration) : base(dbContextOptions)
-        {
-            Configuration = configuration;
-        }
-
-        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-        {
-            //if (!optionsBuilder.IsConfigured)
-            //    base.OnConfiguring(
-            //        optionsBuilder.UseSqlServer(Configuration.GetConnectionString("SomeConnectionString")));
-        }
-
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-            modelBuilder.Entity<SomeFeatureEntity>(a =>
-            {
-                a.ToTable("SomeFeatureEntities").HasKey(k => k.Id);
-                a.Property(p => p.Id).HasColumnName("Id");
-                a.Property(p => p.Name).HasColumnName("Name");
-            });
-
-
-
-            SomeFeatureEntity[] someFeatureEntitySeeds = { new(1, "Test 1"), new(2, "Test 2") };
-            modelBuilder.Entity<SomeFeatureEntity>().HasData(someFeatureEntitySeeds);
-
-           
-        }
-    }
-}
