--- src/demoProjects/templateProject/Persistence/PersistenceServiceRegistration.cs
@@ -1,28 +1,0 @@
-﻿using Application.Services.Repositories;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Persistence.Contexts;
-using Persistence.Repositories;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Persistence
-{
-    public static class PersistenceServiceRegistration
-    {
-        public static IServiceCollection AddPersistenceServices(this IServiceCollection services,
-                                                                IConfiguration configuration)
-        {
-            services.AddDbContext<BaseDbContext>(options =>
-                                                     options.UseSqlServer(
-                                                         configuration.GetConnectionString("TemplateProjectConnectionString")));
-            services.AddScoped<ISomeFeatureEntityRepository, SomeFeatureEntityRepository>();
-
-            return services;
-        }
-    }
-}
