--- src/demoProjects/templateProject/WebAPI/Controllers/BaseController.cs
@@ -1,13 +1,0 @@
-﻿using MediatR;
-using Microsoft.AspNetCore.Mvc;
-
-namespace WebAPI.Controllers
-{
-    public class BaseController : ControllerBase
-    {
-        protected IMediator? Mediator => _mediator ??= HttpContext.RequestServices.GetService<IMediator>();
-        private IMediator? _mediator;
-
-        
-    }
-}
