--- src/demoProjects/templateProject/WebAPI/Program.cs
@@ -1,42 +1,0 @@
-using Application;
-using Core.CrossCuttingConcerns.Exceptions;
-using Infrastructure;
-using Microsoft.EntityFrameworkCore;
-using Persistence;
-using Persistence.Contexts;
-
-var builder = WebApplication.CreateBuilder(args);
-
-// Add services to the container.
-
-builder.Services.AddControllers();
-builder.Services.AddApplicationServices();
-builder.Services.AddSecurityServices();
-builder.Services.AddPersistenceServices(builder.Configuration);
-builder.Services.AddInfrastructureServices();
-builder.Services.AddHttpContextAccessor();
-
-
-
-builder.Services.AddControllers();
-// Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
-builder.Services.AddEndpointsApiExplorer();
-builder.Services.AddSwaggerGen();
-
-var app = builder.Build();
-
-// Configure the HTTP request pipeline.
-if (app.Environment.IsDevelopment())
-{
-    app.UseSwagger();
-    app.UseSwaggerUI();
-}
-
-if (app.Environment.IsProduction())
-    app.ConfigureCustomExceptionMiddleware();
-
-app.UseAuthorization();
-
-app.MapControllers();
-
-app.Run();
