--- src/demoProjects/templateProject/WebAPI/appsettings.json
@@ -1,12 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "ConnectionStrings": {
-    "TemplateProjectConnectionString": "Server=(localdb)\\MSSQLLocalDB;Database=TemplateProjectDb; Trusted_Connection=True;"
-  },
-  "AllowedHosts": "*"
-}
